main: Move init to own file
